requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
pillow
pandas
openpyxl
python-pptx
tqdm